restructure headers, add more utils

+++ src/include/time.h
@@ -5,6 +5,6 @@
 
 void tick_generator_init(void);
 uint64_t current_us(void);
-void delay_us(uint32_t);
+void delay_us(uint64_t);
 
 #endif
